chore: Unify examples
