+++ Cargo.toml
@@ -19,15 +19,17 @@ lto = false
 opt-level = 3
 
 [dependencies]
-hal = { package = "esp32-hal", version = "0.13.0" }
-esp-backtrace = { version = "0.6.0", features = ["esp32", "panic-handler", "print-uart", "exception-handler"] }
-esp-println       = { version = "0.4.0", features = ["esp32","log"] }
+esp32-hal = { version = "0.13.0", optional = true }
+esp32c3-hal = { version = "0.10.0", optional = true }
+esp32s3-hal = { version = "0.10.0", optional = true }
+esp-backtrace = { version = "0.6.0", features = ["panic-handler", "print-uart", "exception-handler"] }
+esp-println       = { version = "0.4.0", features = ["log"] }
 
 embassy-time = { version = "0.1.0", features = ["nightly"], optional = true }
-embassy-executor = { version = "0.2.0", package = "embassy-executor", features = ["nightly", "executor-thread", "integrated-timers", "arch-xtensa"], optional = true }
+embassy-executor = { version = "0.2.0", package = "embassy-executor", features = ["nightly", "executor-thread", "integrated-timers"], optional = true }
 embassy-net = { version = "0.1.0", features = ["nightly", "tcp", "udp", "dhcpv4", "medium-ethernet"], optional = true }
 
-esp-wifi = { git = "https://github.com/esp-rs/esp-wifi.git", features = ["big-heap", "phy-enable-usb", "esp32", "embedded-svc", "wifi"], rev = "68dc11bbb2c0efa29c4acbbf134d6f142441065e" }
+esp-wifi = { git = "https://github.com/esp-rs/esp-wifi.git", features = ["big-heap", "phy-enable-usb", "embedded-svc", "wifi"], rev = "68dc11bbb2c0efa29c4acbbf134d6f142441065e" }
 smoltcp = { version = "0.10.0", default-features=false, features = ["proto-igmp", "proto-ipv4", "socket-tcp", "socket-icmp", "socket-udp", "medium-ethernet", "proto-dhcpv4", "socket-raw", "socket-dhcpv4"] }
 embedded-svc = { version = "0.25.0", default-features = false}
 log = "0.4.16"
@@ -36,7 +38,9 @@ embedded-io = "0.4.0"
 heapless = "0.7.16"
 critical-section = "1.0.1"
 
-esp-mbedtls = { path = "../esp-mbedtls", features = ["async"] }
+esp-mbedtls = { path = "./esp-mbedtls" }
+
+[target.xtensa-esp32s3-none-elf.dependencies]
 
 [[example]]
 name              = "async_client"
@@ -55,7 +59,10 @@ name              = "async_server_mTLS"
 required-features = ["async"]
 
 [features]
-async = ["esp-wifi/async", "esp-wifi/embassy-net", "embassy-executor", "embassy-net", "embassy-time", "embedded-io/async", "hal/embassy-time-timg0"]
+esp32 = ["esp32-hal/embassy-time-timg0", "esp-backtrace/esp32", "esp-println/esp32", "embassy-executor?/arch-xtensa", "esp-wifi/esp32"]
+esp32c3 = ["esp32c3-hal/embassy-time-timg0", "esp-backtrace/esp32c3", "esp-println/esp32c3", "embassy-executor?/arch-riscv32", "esp-wifi/esp32c3"]
+esp32s3 = ["esp32s3-hal/embassy-time-timg0", "esp-backtrace/esp32s3", "esp-println/esp32s3", "embassy-executor?/arch-xtensa", "esp-wifi/esp32s3"]
+async = ["esp-wifi/async", "esp-wifi/embassy-net", "embassy-executor", "embassy-net", "embassy-time", "embedded-io/async", "esp-mbedtls/async"]
 # Use this feature to enable the private key with a password
 encrypted_private_key = []
