+++ README.md
@@ -21,17 +21,50 @@ Also the examples include hard-coded certificates which will expire some day.
 
 ## Running Examples
 
-`cd` into the directory, e.g. `examples-esp32c3`, then use
+Examples are available for:
+
+- esp32
+- esp32c3
+- esp32s3
+
+To run examples, you need to specify the architecture as a feature, the example name, the target and the toolchain.
+
+You also need to set `SSID` and `PASSWORD` as your environment variables
+
+### Examples
+
+Xtensa:
 
 ```shell
-cargo run --release --example sync
+SSID=<your_ssid> PASSWORD=<your_password> cargo +esp run --release --example sync_client -F esp32s3 --target xtensa-esp32s3-none-elf
 ```
 
-or
+RISC-V: 
 
 ```shell
-cargo run --release --example async --features=async
+SSID=<your_ssid> PASSWORD=<your_password> cargo +nightly-2023-03-09 run --release --example async_client -F esp32c3,async --target riscv32imc-unknown-none-elf
 ```
+
+Here's a table of the architectures with their corresponding target for quick reference:
+
+| Architecture | Target                      | Toolchain          |
+| ------------ | --------------------------- | ------------------ |
+| esp32        | xtensa-esp32-none-elf       | esp                |
+| esp32c3      | riscv32imc-unknown-none-elf | nightly-2023-03-09 |
+| esp32s3      | xtensa-esp32s3-none-elf     | esp                |
+
+Heres's a list of all the examples with their description:
+
+| Example           | Description                                                  |
+| :---------------- | ------------------------------------------------------------ |
+| async_client      | Example of a HTTPS connection using the async client.        |
+| async_client_mTLS | Example of a HTTPS connection using the async client, with certificate authentication. This sends client certificates to a server, and the response indicates informations about the certificates. |
+| sync_client       | Example of a HTTPS connection using the sync client.         |
+| sync_client_mTLS  | Example of a HTTPS connection using the sync client, with certificate authentication. This sends client certificates to a server, and the response indicates informations about the certificates. |
+| async_server      | Example of a simple async server with HTTPS support. This uses self-signed certificates, so you will need to enable an exception in your browser. |
+| async_server_mTLS | Example of a simple async server with HTTPS support, with client authentication. You will need to pass client certificates in your request in order to have a successful connection. Refer to the documentation inside the example. |
+| sync_server       | Example of a simple sync server with HTTPS support. This uses self-signed certificates, so you will need to enable an exception in your browser. |
+| sync_server_mTLS  | Example of a simple sync server with HTTPS support, with client authentication. You will need to pass client certificates in your request in order to have a successful connection. Refer to the documentation inside the example. |
 
 This needs `espflash` version 2.x. If you are using version 1.x you need to remove the `flash` command from the runner in `.cargo/config.toml`
