--- examples-esp32/.cargo/config.toml
@@ -1,15 +1,0 @@
-[target.xtensa-esp32-none-elf]
-runner = "espflash flash --monitor"
-
-[build]
-rustflags = [
-  "-C", "link-arg=-Tlinkall.x",
-  "-C", "link-arg=-nostartfiles",
-  "-C", "link-arg=-Trom_functions.x",
-  "-C", "target-feature=-loop",
-]
-
-target = "xtensa-esp32-none-elf"
-
-[unstable]
-build-std = ["core"]
