--- examples-esp32c3/src/lib.rs
@@ -1,3 +1,0 @@
-#![no_std]
-
-// see examples
