--- examples-esp32s3/Cargo.toml
@@ -1,61 +1,0 @@
-[package]
-name = "examples"
-version = "0.1.0"
-authors = ["bjoernQ <bjoern.quentin@mobile-j.de>"]
-edition = "2021"
-license = "MIT OR Apache-2.0"
-
-[profile.release]
-debug = true
-lto = false
-
-[profile.release.package.esp-wifi]
-opt-level = 3
-
-[profile.dev]
-lto = false
-
-[profile.dev.package.esp-wifi]
-opt-level = 3
-
-[dependencies]
-hal = { package = "esp32s3-hal", version = "0.10.0" }
-esp-backtrace = { version = "0.6.0", features = ["esp32s3", "panic-handler", "print-uart", "exception-handler"] }
-esp-println       = { version = "0.4.0", features = ["esp32s3","log"] }
-
-embassy-time = { version = "0.1.0", features = ["nightly"], optional = true }
-embassy-executor = { version = "0.2.0", package = "embassy-executor", features = ["nightly", "executor-thread", "integrated-timers", "arch-xtensa"], optional = true }
-embassy-net = { version = "0.1.0", features = ["nightly", "tcp", "udp", "dhcpv4", "medium-ethernet"], optional = true }
-
-esp-wifi = { git = "https://github.com/esp-rs/esp-wifi.git", features = ["big-heap", "phy-enable-usb", "esp32s3", "embedded-svc", "wifi"], rev = "68dc11bbb2c0efa29c4acbbf134d6f142441065e" }
-smoltcp = { version = "0.10.0", default-features=false, features = ["proto-igmp", "proto-ipv4", "socket-tcp", "socket-icmp", "socket-udp", "medium-ethernet", "proto-dhcpv4", "socket-raw", "socket-dhcpv4"] }
-embedded-svc = { version = "0.25.0", default-features = false}
-log = "0.4.16"
-embedded-hal = "0.2"
-embedded-io = "0.4.0"
-heapless = "0.7.16"
-critical-section = "1.0.1"
-
-esp-mbedtls = { path = "../esp-mbedtls", features = ["async"] }
-
-[[example]]
-name              = "async_client"
-required-features = ["async"]
-
-[[example]]
-name              = "async_client_mTLS"
-required-features = ["async"]
-
-[[example]]
-name              = "async_server"
-required-features = ["async"]
-
-[[example]]
-name              = "async_server_mTLS"
-required-features = ["async"]
-
-[features]
-async = ["esp-wifi/async", "esp-wifi/embassy-net", "embassy-executor", "embassy-net", "embassy-time", "embedded-io/async", "hal/embassy-time-timg0"]
-# Use this feature to enable the private key with a password
-encrypted_private_key = []
-
