--- examples-esp32s3/rust-toolchain.toml
@@ -1,2 +1,0 @@
-[toolchain]
-channel = "esp"
