+++ examples/async_server_mTLS.rs
@@ -1,4 +1,4 @@
-//! Example for a sync server.
+//! Example for an async server.
 //! Contains a basic server implementation to test mbedtls in server mode.
 //!
 //! This example is configured to use mTLS. If you send a request, without passing
@@ -99,151 +99,237 @@
 //! ```
 #![no_std]
 #![no_main]
+#![feature(type_alias_impl_trait)]
 #![allow(non_snake_case)]
 
-use embedded_io::blocking::*;
-use embedded_svc::{
-    ipv4::Interface,
-    wifi::{ClientConfiguration, Configuration, Wifi},
-};
+#[doc(hidden)]
+#[cfg(feature = "esp32")]
+pub use esp32_hal as hal;
+#[doc(hidden)]
+#[cfg(feature = "esp32c3")]
+pub use esp32c3_hal as hal;
+#[doc(hidden)]
+#[cfg(feature = "esp32s3")]
+pub use esp32s3_hal as hal;
+
+use embassy_executor::_export::StaticCell;
+use embassy_net::tcp::TcpSocket;
+use embassy_net::{Config, Stack, StackResources};
+
+use embassy_executor::Executor;
+use embassy_time::{Duration, Timer};
+use embedded_svc::wifi::{ClientConfiguration, Configuration, Wifi};
 use esp_backtrace as _;
-use esp_mbedtls::{set_debug, Mode, TlsError, TlsVersion};
-use esp_mbedtls::{Certificates, Session, X509};
-use esp_println::{logger::init_logger, print, println};
-use esp_wifi::{
-    current_millis,
-    wifi::{utils::create_network_interface, WifiMode},
-    wifi_interface::WifiStack,
-    EspWifiInitFor,
-};
-use hal::timer::TimerGroup;
-use hal::{
-    clock::{ClockControl, CpuClock},
-    peripherals::Peripherals,
-    prelude::*,
-    Rng, Rtc,
-};
-use smoltcp::iface::SocketStorage;
+use esp_mbedtls::{asynch::Session, set_debug, Certificates, Mode, TlsVersion};
+use esp_mbedtls::{TlsError, X509};
+use esp_println::logger::init_logger;
+use esp_println::{print, println};
+use esp_wifi::wifi::{WifiController, WifiDevice, WifiEvent, WifiMode, WifiState};
+use esp_wifi::{initialize, EspWifiInitFor};
+use hal::clock::{ClockControl, CpuClock};
+use hal::Rng;
+use hal::{embassy, peripherals::Peripherals, prelude::*, timer::TimerGroup, Rtc};
+use smoltcp::wire::IpListenEndpoint;
 
 const SSID: &str = env!("SSID");
 const PASSWORD: &str = env!("PASSWORD");
+
+macro_rules! singleton {
+    ($val:expr) => {{
+        type T = impl Sized;
+        static STATIC_CELL: StaticCell<T> = StaticCell::new();
+        let (x,) = STATIC_CELL.init(($val,));
+        x
+    }};
+}
+
+static EXECUTOR: StaticCell<Executor> = StaticCell::new();
 
 #[entry]
 fn main() -> ! {
     init_logger(log::LevelFilter::Info);
 
     let peripherals = Peripherals::take();
+    #[cfg(feature = "esp32")]
     let mut system = peripherals.DPORT.split();
+    #[cfg(not(feature = "esp32"))]
+    #[allow(unused_mut)]
+    let mut system = peripherals.SYSTEM.split();
+    #[cfg(feature = "esp32c3")]
+    let clocks = ClockControl::configure(system.clock_control, CpuClock::Clock160MHz).freeze();
+    #[cfg(any(feature = "esp32", feature = "esp32s3"))]
     let clocks = ClockControl::configure(system.clock_control, CpuClock::Clock240MHz).freeze();
 
     let mut rtc = Rtc::new(peripherals.RTC_CNTL);
 
     // Disable watchdog timers
+    #[cfg(not(feature = "esp32"))]
+    rtc.swd.disable();
     rtc.rwdt.disable();
 
-    let rngp = Rng::new(peripherals.RNG);
-    let timer = TimerGroup::new(
+    #[cfg(feature = "esp32c3")]
+    let timer = hal::systimer::SystemTimer::new(peripherals.SYSTIMER).alarm0;
+    #[cfg(any(feature = "esp32", feature = "esp32s3"))]
+    let timer = hal::timer::TimerGroup::new(
         peripherals.TIMG1,
         &clocks,
         &mut system.peripheral_clock_control,
-    );
-    let init = esp_wifi::initialize(
+    )
+    .timer0;
+    let init = initialize(
         EspWifiInitFor::Wifi,
-        timer.timer0,
-        rngp,
+        timer,
+        Rng::new(peripherals.RNG),
         system.radio_clock_control,
         &clocks,
     )
     .unwrap();
 
     let (wifi, _) = peripherals.RADIO.split();
-    let mut socket_set_entries: [SocketStorage; 3] = Default::default();
-    let (iface, device, mut controller, sockets) =
-        create_network_interface(&init, wifi, WifiMode::Sta, &mut socket_set_entries);
-    let wifi_stack = WifiStack::new(iface, device, sockets, current_millis);
+    let (wifi_interface, controller) = esp_wifi::wifi::new_with_mode(&init, wifi, WifiMode::Sta);
 
-    println!("Call wifi_connect");
-    let client_config = Configuration::Client(ClientConfiguration {
-        ssid: SSID.into(),
-        password: PASSWORD.into(),
-        ..Default::default()
+    let timer_group0 = TimerGroup::new(
+        peripherals.TIMG0,
+        &clocks,
+        &mut system.peripheral_clock_control,
+    );
+    embassy::init(&clocks, timer_group0.timer0);
+
+    let config = Config::dhcpv4(Default::default());
+
+    let seed = 1234; // very random, very secure seed
+
+    // Init network stack
+    let stack = &*singleton!(Stack::new(
+        wifi_interface,
+        config,
+        singleton!(StackResources::<3>::new()),
+        seed
+    ));
+
+    let executor = EXECUTOR.init(Executor::new());
+    executor.run(|spawner| {
+        spawner.spawn(connection(controller)).ok();
+        spawner.spawn(net_task(&stack)).ok();
+        spawner.spawn(task(&stack)).ok();
     });
-    controller.set_configuration(&client_config).unwrap();
-    controller.start().unwrap();
-    controller.connect().unwrap();
+}
 
-    println!("Wait to get connected");
+#[embassy_executor::task]
+async fn connection(mut controller: WifiController<'static>) {
+    println!("start connection task");
+    println!("Device capabilities: {:?}", controller.get_capabilities());
     loop {
-        let res = controller.is_connected();
-        match res {
-            Ok(connected) => {
-                if connected {
-                    break;
-                }
+        match esp_wifi::wifi::get_wifi_state() {
+            WifiState::StaConnected => {
+                // wait until we're no longer connected
+                controller.wait_for_event(WifiEvent::StaDisconnected).await;
+                Timer::after(Duration::from_millis(5000)).await
             }
-            Err(err) => {
-                println!("{:?}", err);
-                loop {}
+            _ => {}
+        }
+        if !matches!(controller.is_started(), Ok(true)) {
+            let client_config = Configuration::Client(ClientConfiguration {
+                ssid: SSID.into(),
+                password: PASSWORD.into(),
+                ..Default::default()
+            });
+            controller.set_configuration(&client_config).unwrap();
+            println!("Starting wifi");
+            controller.start().await.unwrap();
+            println!("Wifi started!");
+        }
+        println!("About to connect...");
+
+        match controller.connect().await {
+            Ok(_) => println!("Wifi connected!"),
+            Err(e) => {
+                println!("Failed to connect to wifi: {e:?}");
+                Timer::after(Duration::from_millis(5000)).await
             }
         }
     }
+}
 
-    // wait for getting an ip address
-    println!("Wait to get an ip address");
+#[embassy_executor::task]
+async fn net_task(stack: &'static Stack<WifiDevice<'static>>) {
+    stack.run().await
+}
+
+#[embassy_executor::task]
+async fn task(stack: &'static Stack<WifiDevice<'static>>) {
+    let mut rx_buffer = [0; 4096];
+    let mut tx_buffer = [0; 4096];
+
     loop {
-        wifi_stack.work();
-
-        if wifi_stack.is_iface_up() {
-            println!("Got ip {:?}", wifi_stack.get_ip_info());
+        if stack.is_link_up() {
             break;
         }
+        Timer::after(Duration::from_millis(500)).await;
     }
 
-    println!("We are connected!");
-
-    println!(
-        "Point your browser to https://{:?}/",
-        wifi_stack.get_ip_info().unwrap().ip
-    );
-    let mut rx_buffer = [0u8; 1536];
-    let mut tx_buffer = [0u8; 1536];
-    let mut socket = wifi_stack.get_socket(&mut rx_buffer, &mut tx_buffer);
-
-    socket.listen(443).unwrap();
-    set_debug(0);
+    println!("Waiting to get IP address...");
     loop {
-        socket.work();
+        if let Some(config) = stack.config_v4() {
+            println!("Got IP: {}", config.address);
+            println!(
+                "Point your browser to https://{}/",
+                config.address.address()
+            );
+            break;
+        }
+        Timer::after(Duration::from_millis(500)).await;
+    }
 
-        if !socket.is_open() {
-            socket.listen(443).unwrap();
+    let mut socket = TcpSocket::new(&stack, &mut rx_buffer, &mut tx_buffer);
+    socket.set_timeout(Some(Duration::from_secs(10)));
+    loop {
+        println!("Waiting for connection...");
+        let r = socket
+            .accept(IpListenEndpoint {
+                addr: None,
+                port: 443,
+            })
+            .await;
+        println!("Connected...");
+
+        if let Err(e) = r {
+            println!("connect error: {:?}", e);
+            continue;
         }
 
-        if socket.is_connected() {
-            println!("New connection");
+        set_debug(0);
+        use embedded_io::asynch::Read;
+        use embedded_io::asynch::Write;
 
-            let mut time_out = false;
-            let wait_end = current_millis() + 20 * 1000;
-            let mut buffer = [0u8; 1024];
-            let mut pos = 0;
+        let mut buffer = [0u8; 1024];
+        let mut pos = 0;
+        let tls: Session<_, 4096> = Session::new(
+            &mut socket,
+            "",
+            Mode::Server,
+            TlsVersion::Tls1_2,
+            Certificates {
+                ca_chain: X509::pem(CA_CERT.as_bytes()).ok(),
+                // Use self-signed certificates
+                certificate: X509::pem(CERT.as_bytes()).ok(),
+                private_key: X509::pem(PRIVATE_KEY.as_bytes()).ok(),
+                ..Default::default()
+            },
+        )
+        .unwrap();
 
-            let tls = Session::new(
-                &mut socket,
-                "",
-                Mode::Server,
-                TlsVersion::Tls1_2,
-                Certificates {
-                    ca_chain: X509::pem(CA_CERT.as_bytes()).ok(),
-                    // Use self-signed certificates
-                    certificate: X509::pem(CERT.as_bytes()).ok(),
-                    private_key: X509::pem(PRIVATE_KEY.as_bytes()).ok(),
-                    ..Default::default()
-                },
-            )
-            .unwrap();
-            match tls.connect() {
-                Ok(mut connected_session) => {
-                    loop {
-                        if let Ok(len) = connected_session.read(&mut buffer[pos..]) {
+        println!("Start tls connect");
+        match tls.connect().await {
+            Ok(mut connected_session) => {
+                log::info!("Got session");
+                loop {
+                    match connected_session.read(&mut buffer).await {
+                        Ok(0) => {
+                            println!("read EOF");
+                            break;
+                        }
+                        Ok(len) => {
                             let to_print =
                                 unsafe { core::str::from_utf8_unchecked(&buffer[..(pos + len)]) };
 
@@ -254,56 +340,48 @@ fn main() -> ! {
                             }
 
                             pos += len;
-                        } else {
+                        }
+                        Err(e) => {
+                            println!("read error: {:?}", e);
                             break;
                         }
+                    };
+                }
 
-                        if current_millis() > wait_end {
-                            println!("Timed out");
-                            time_out = true;
-                            break;
-                        }
-                    }
-
-                    if !time_out {
-                        connected_session
-                            .write_all(
-                                b"HTTP/1.0 200 OK\r\n\r\n\
-                                    <html>\
-                                    <body>\
+                let r = connected_session
+                    .write_all(
+                        b"HTTP/1.0 200 OK\r\n\r\n\
+                            <html>\
+                                <body>\
                                     <h1>Hello Rust! Hello esp-mbedtls!</h1>\
-                                    </body>\
-                                    </html>\r\n\
-                                    ",
-                            )
-                            .unwrap();
-                    }
+                                </body>\
+                            </html>\r\n\
+                            ",
+                    )
+                    .await;
+                if let Err(e) = r {
+                    println!("write error: {:?}", e);
+                }
 
-                    drop(connected_session);
-                }
-                Err(TlsError::NoClientCertificate) => {
-                    println!("Error: No client certificates given. Please provide client certificates during your request");
-                }
-                Err(TlsError::MbedTlsError(-30592)) => {
-                    println!("Fatal message: Please enable the exception for a self-signed certificate in your browser");
-                }
-                Err(error) => {
-                    panic!("{:?}", error);
-                }
+                Timer::after(Duration::from_millis(1000)).await;
+
+                drop(connected_session);
             }
-
-            socket.close();
-
-            println!("Done\n");
-            println!();
+            Err(TlsError::NoClientCertificate) => {
+                println!("Error: No client certificates given. Please provide client certificates during your request");
+            }
+            Err(TlsError::MbedTlsError(-30592)) => {
+                println!("Fatal message: Please enable the exception for a self-signed certificate in your browser");
+            }
+            Err(error) => {
+                panic!("{:?}", error);
+            }
         }
+        println!("Closing socket");
+        socket.close();
+        Timer::after(Duration::from_millis(1000)).await;
 
-        // This seems to delay after a connection. Removed to allow instant connections
-        //
-        // let wait_end = current_millis() + 5 * 1000;
-        // while current_millis() < wait_end {
-        //     socket.work();
-        // }
+        socket.abort();
     }
 }
